# gen/http/openapi.json
{"swagger":"2.0","info":{"title":"Feed Generator","description":"Generates feed for different types of media","version":""},"host":"localhost:80","consumes":["application/json","application/xml","application/gob"],"produces":["application/json","application/xml","application/gob"],"paths":{"/feed/manga":{"post":{"tags":["feedgen"],"summary":"manga feedgen","operationId":"feedgen#manga","parameters":[{"name":"feedType","in":"query","description":"RSS, Atom, or JSON Feed","required":false,"type":"string","default":"json","enum":["rss","atom","json"]},{"name":"MangaRequestBody","in":"body","required":true,"schema":{"$ref":"#/definitions/FeedgenMangaRequestBody","required":["titles"]}}],"responses":{"200":{"description":"OK response.","schema":{"type":"string"}},"404":{"description":"Not Found response.","schema":{"$ref":"#/definitions/Feedgenmanga_NotFound_response_body"}},"500":{"description":"Internal Server Error response.","schema":{"$ref":"#/definitions/Feedgenmanga_InternalServerError_response_body"}},"502":{"description":"Bad Gateway response.","schema":{"$ref":"#/definitions/Feedgenmanga_BadGateway_response_body"}}},"schemes":["http"]}},"/feed/manga/{hash}":{"get":{"tags":["feedgen"],"summary":"viewManga feedgen","operationId":"feedgen#viewManga","parameters":[{"name":"hash","in":"path","description":"Identifier of previously created manga feed","required":true,"type":"string"}],"responses":{"200":{"description":"OK response.","schema":{"type":"string","format":"byte"},"headers":{"Content-Type":{"type":"string"}}},"404":{"description":"Not Found response.","schema":{"$ref":"#/definitions/FeedgenviewManga_NotFound_response_body"}},"500":{"description":"Internal Server Error response.","schema":{"$ref":"#/definitions/FeedgenviewManga_InternalServerError_response_body"}},"502":{"description":"Bad Gateway response.","schema":{"$ref":"#/definitions/FeedgenviewManga_BadGateway_response_body"}}},"schemes":["http"]}}},"definitions":{"FeedgenMangaRequestBody":{"title":"FeedgenMangaRequestBody","type":"object","properties":{"titles":{"type":"array","items":{"type":"string","example":"Labore dolor qui."},"description":"List of manga titles to subscribe to","example":["Rem et aliquam non.","Velit temporibus qui voluptatem sunt."],"minItems":1,"maxItems":2048}},"example":{"titles":["Quasi qui sit et magni suscipit."]},"required":["titles"]},"Feedgenmanga_BadGateway_response_body":{"title":"Mediatype identifier: application/vnd.goa.error; view=default","type":"object","properties":{"fault":{"type":"boolean","description":"Is the error a server-side fault?","example":true},"id":{"type":"string","description":"ID is a unique identifier for this particular occurrence of the problem.","example":"123abc"},"message":{"type":"string","description":"Message is a human-readable explanation specific to this occurrence of the problem.","example":"parameter 'p' must be an integer"},"name":{"type":"string","description":"Name is the name of this class of errors.","example":"bad_request"},"temporary":{"type":"boolean","description":"Is the error temporary?","example":true},"timeout":{"type":"boolean","description":"Is the error a timeout?","example":true}},"description":"manga_BadGateway_response_body result type (default view)","example":{"fault":true,"id":"123abc","message":"parameter 'p' must be an integer","name":"bad_request","temporary":true,"timeout":false},"required":["name","id","message","temporary","timeout","fault"]},"Feedgenmanga_InternalServerError_response_body":{"title":"Mediatype identifier: application/vnd.goa.error; view=default","type":"object","properties":{"fault":{"type":"boolean","description":"Is the error a server-side fault?","example":true},"id":{"type":"string","description":"ID is a unique identifier for this particular occurrence of the problem.","example":"123abc"},"message":{"type":"string","description":"Message is a human-readable explanation specific to this occurrence of the problem.","example":"parameter 'p' must be an integer"},"name":{"type":"string","description":"Name is the name of this class of errors.","example":"bad_request"},"temporary":{"type":"boolean","description":"Is the error temporary?","example":true},"timeout":{"type":"boolean","description":"Is the error a timeout?","example":false}},"description":"manga_InternalServerError_response_body result type (default view)","example":{"fault":false,"id":"123abc","message":"parameter 'p' must be an integer","name":"bad_request","temporary":false,"timeout":true},"required":["name","id","message","temporary","timeout","fault"]},"Feedgenmanga_NotFound_response_body":{"title":"Mediatype identifier: application/vnd.goa.error; view=default","type":"object","properties":{"fault":{"type":"boolean","description":"Is the error a server-side fault?","example":true},"id":{"type":"string","description":"ID is a unique identifier for this particular occurrence of the problem.","example":"123abc"},"message":{"type":"string","description":"Message is a human-readable explanation specific to this occurrence of the problem.","example":"parameter 'p' must be an integer"},"name":{"type":"string","description":"Name is the name of this class of errors.","example":"bad_request"},"temporary":{"type":"boolean","description":"Is the error temporary?","example":false},"timeout":{"type":"boolean","description":"Is the error a timeout?","example":false}},"description":"manga_NotFound_response_body result type (default view)","example":{"fault":false,"id":"123abc","message":"parameter 'p' must be an integer","name":"bad_request","temporary":true,"timeout":false},"required":["name","id","message","temporary","timeout","fault"]},"FeedgenviewManga_BadGateway_response_body":{"title":"Mediatype identifier: application/vnd.goa.error; view=default","type":"object","properties":{"fault":{"type":"boolean","description":"Is the error a server-side fault?","example":true},"id":{"type":"string","description":"ID is a unique identifier for this particular occurrence of the problem.","example":"123abc"},"message":{"type":"string","description":"Message is a human-readable explanation specific to this occurrence of the problem.","example":"parameter 'p' must be an integer"},"name":{"type":"string","description":"Name is the name of this class of errors.","example":"bad_request"},"temporary":{"type":"boolean","description":"Is the error temporary?","example":true},"timeout":{"type":"boolean","description":"Is the error a timeout?","example":true}},"description":"viewManga_BadGateway_response_body result type (default view)","example":{"fault":true,"id":"123abc","message":"parameter 'p' must be an integer","name":"bad_request","temporary":false,"timeout":false},"required":["name","id","message","temporary","timeout","fault"]},"FeedgenviewManga_InternalServerError_response_body":{"title":"Mediatype identifier: application/vnd.goa.error; view=default","type":"object","properties":{"fault":{"type":"boolean","description":"Is the error a server-side fault?","example":false},"id":{"type":"string","description":"ID is a unique identifier for this particular occurrence of the problem.","example":"123abc"},"message":{"type":"string","description":"Message is a human-readable explanation specific to this occurrence of the problem.","example":"parameter 'p' must be an integer"},"name":{"type":"string","description":"Name is the name of this class of errors.","example":"bad_request"},"temporary":{"type":"boolean","description":"Is the error temporary?","example":true},"timeout":{"type":"boolean","description":"Is the error a timeout?","example":true}},"description":"viewManga_InternalServerError_response_body result type (default view)","example":{"fault":false,"id":"123abc","message":"parameter 'p' must be an integer","name":"bad_request","temporary":true,"timeout":false},"required":["name","id","message","temporary","timeout","fault"]},"FeedgenviewManga_NotFound_response_body":{"title":"Mediatype identifier: application/vnd.goa.error; view=default","type":"object","properties":{"fault":{"type":"boolean","description":"Is the error a server-side fault?","example":true},"id":{"type":"string","description":"ID is a unique identifier for this particular occurrence of the problem.","example":"123abc"},"message":{"type":"string","description":"Message is a human-readable explanation specific to this occurrence of the problem.","example":"parameter 'p' must be an integer"},"name":{"type":"string","description":"Name is the name of this class of errors.","example":"bad_request"},"temporary":{"type":"boolean","description":"Is the error temporary?","example":true},"timeout":{"type":"boolean","description":"Is the error a timeout?","example":false}},"description":"viewManga_NotFound_response_body result type (default view)","example":{"fault":true,"id":"123abc","message":"parameter 'p' must be an integer","name":"bad_request","temporary":false,"timeout":false},"required":["name","id","message","temporary","timeout","fault"]}}}